test(utils): createReactApp moved from server to mocked
